// frontend/src/types/PostPartrum.ts
export interface PostPartrumModel {
    keyId?: string;
    postPartrumId?: number;
    newBornId?: number;
    postPartrumDate?: string;
    bodyTemperature?: string;
    postPartrumBP?: string;
    postPartrumFundus?: string;
    breast?: string;
    Lochia?: string;
    postPartrumRemarks: string;
}